# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Desafio Backend Pleno - Soluct API",
        "description": "Documentação da API RESTful para o desafio técnico da Soluct.\n * * **Observações Importantes:**\n * - Todas as requisições que enviam dados no corpo (POST, PUT, PATCH) devem usar o header `Content-Type: application/json`.\n * - Todas as requisições devem incluir o header `Accept: application/json` para garantir que a resposta seja recebida no formato JSON correto.\n * - Os endpoints que requerem autenticação estão marcados com um ícone de cadeado.",
        "contact": {
            "email": "elizaelnycolas@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/me": {
            "get": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Obter informações do usuário autenticado",
                "description": "Retorna os dados do usuário atualmente autenticado.",
                "operationId": "f5719ff43b3749e6a1457db1b25ddab5",
                "responses": {
                    "200": {
                        "description": "Dados do usuário autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Nicolas Araujo"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "nicolas@example.com"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-06-25T20:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-06-25T20:30:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Usuário não autenticado"
                    },
                    "500": {
                        "description": "Erro interno ao buscar usuário"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Atualizar dados do usuário autenticado",
                "description": "Atualiza os dados do usuário autenticado. Permite alterar nome, email e senha.",
                "operationId": "ffd761d16bd5323e0f03f82229476b93",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Nome do Usuário"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255,
                                        "example": "usuario@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "minLength": 8,
                                        "example": "novaSenha123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "minLength": 8,
                                        "example": "novaSenha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário atualizado com sucesso."
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Nome do Usuário"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "usuario@email.com"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-06-25T20:00:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-06-25T20:30:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro de validação."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao atualizar usuário"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Excluir conta do usuário autenticado",
                "description": "Exclui a conta do usuário autenticado e revoga todos os tokens.",
                "operationId": "38e57d6e299f9699c9f90eca2a341459",
                "responses": {
                    "200": {
                        "description": "Conta excluída com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Conta excluída com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Usuário não autenticado"
                    },
                    "500": {
                        "description": "Erro interno ao excluir conta"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
...
}